test(azure): isolate redirect security coverage
