+++ tests/lib/azure.test.ts
@@ -1,6 +1,5 @@
-import { createServer } from 'node:http';
 import { vi } from 'vitest';
-import { AzureOpenAI, APIError, APIUserAbortError, toStreamingFile } from 'openai';
+import { AzureOpenAI, APIUserAbortError, toStreamingFile } from 'openai';
 import type { AzureClientOptions } from 'openai';
 import type { RequestInit, RequestInfo, Response } from 'openai/internal/builtin-types';
 
@@ -365,157 +364,6 @@ describe('instantiate azure client', () => {
         }),
     ).toThrow(/baseURL and endpoint are mutually exclusive/);
   });
-});
-
-describe('azure redirect safety', () => {
-  const apiKey = 'AZURE_OPENAI_TEST_SECRET';
-  const baseURL = 'http://127.0.0.1';
-
-  test.each([
-    ['default fetch behavior', undefined, undefined],
-    ['client-level redirect override', 'follow', undefined],
-    ['request-level redirect override', undefined, 'follow'],
-    ['client-level and request-level redirect overrides', 'follow', 'follow'],
-  ] as const)(
-    'disables automatic redirects for static API keys despite %s',
-    async (_configuration, clientRedirect, requestRedirect) => {
-      const client = new AzureOpenAI({
-        baseURL,
-        apiKey,
-        apiVersion,
-        ...(clientRedirect ? { fetchOptions: { redirect: clientRedirect } } : {}),
-      });
-
-      const { req } = await client.buildRequest({
-        path: '/foo',
-        method: 'get',
-        ...(requestRedirect ? { fetchOptions: { redirect: requestRedirect } } : {}),
-      });
-
-      expect(req.headers.get('api-key')).toBe(apiKey);
-      expect(req.redirect).toBe('manual');
-    },
-  );
-
-  test.each([
-    ['default redirect behavior', undefined, undefined, undefined],
-    ['client-level redirect preference', 'follow', undefined, 'follow'],
-    ['request-level redirect preference', 'error', 'follow', 'follow'],
-  ] as const)(
-    'preserves %s for bearer-only authentication',
-    async (_configuration, clientRedirect, requestRedirect, expectedRedirect) => {
-      let requestedInit: RequestInit | undefined;
-      const client = new AzureOpenAI({
-        baseURL,
-        azureADTokenProvider: async () => 'azure-ad-token',
-        apiVersion,
-        fetch: async (_url, init) => {
-          requestedInit = init;
-          return Response.json({ ok: true });
-        },
-        ...(clientRedirect ? { fetchOptions: { redirect: clientRedirect } } : {}),
-      });
-
-      await client.get('/foo', requestRedirect ? { fetchOptions: { redirect: requestRedirect } } : undefined);
-
-      const requestedHeaders = new Headers(requestedInit?.headers);
-      expect(requestedHeaders.get('authorization')).toBe('Bearer azure-ad-token');
-      expect(requestedHeaders.has('api-key')).toBe(false);
-      expect(requestedInit?.redirect).toBe(expectedRedirect);
-    },
-  );
-
-  test.each([
-    ['default redirect behavior', undefined, undefined, undefined],
-    ['client-level redirect preference', 'follow', undefined, 'follow'],
-    ['request-level redirect preference', 'error', 'follow', 'follow'],
-  ] as const)(
-    'preserves %s when the API key header is explicitly removed',
-    async (_configuration, clientRedirect, requestRedirect, expectedRedirect) => {
-      const client = new AzureOpenAI({
-        baseURL,
-        apiKey,
-        apiVersion,
-        ...(clientRedirect ? { fetchOptions: { redirect: clientRedirect } } : {}),
-      });
-
-      const { req } = await client.buildRequest({
-        path: '/foo',
-        method: 'get',
-        headers: { 'api-key': null },
-        ...(requestRedirect ? { fetchOptions: { redirect: requestRedirect } } : {}),
-      });
-
-      expect(req.headers.has('api-key')).toBe(false);
-      expect(req.redirect).toBe(expectedRedirect);
-    },
-  );
-
-  test.each([302, 307, 308])(
-    'does not disclose static API keys to another origin on an HTTP %i redirect',
-    async (status) => {
-      const disclosedAPIKeys: (string | string[] | undefined)[] = [];
-      let redirectURL = '';
-
-      const destination = createServer((request, response) => {
-        request.resume();
-        disclosedAPIKeys.push(request.headers['api-key']);
-        response.writeHead(200, { 'content-type': 'application/json' });
-        response.end(JSON.stringify({ leakedApiKey: request.headers['api-key'] }));
-      });
-
-      const source = createServer((request, response) => {
-        request.resume();
-        response.writeHead(status, { location: redirectURL });
-        response.end();
-      });
-
-      try {
-        await Promise.all([
-          new Promise<void>((resolve) => destination.listen(0, '127.0.0.1', resolve)),
-          new Promise<void>((resolve) => source.listen(0, '127.0.0.1', resolve)),
-        ]);
-
-        const destinationAddress = destination.address();
-        const sourceAddress = source.address();
-
-        if (
-          !destinationAddress ||
-          typeof destinationAddress === 'string' ||
-          !sourceAddress ||
-          typeof sourceAddress === 'string'
-        ) {
-          throw new Error('Expected both redirect test servers to bind ephemeral TCP ports');
-        }
-
-        redirectURL = `http://127.0.0.1:${destinationAddress.port}/attacker`;
-
-        const client = new AzureOpenAI({
-          baseURL: `http://127.0.0.1:${sourceAddress.port}`,
-          apiKey,
-          apiVersion,
-          maxRetries: 0,
-          fetchOptions: { redirect: 'follow' },
-        });
-
-        const request = client.get('/redirect', { fetchOptions: { redirect: 'follow' } });
-
-        await expect(request).rejects.toBeInstanceOf(APIError);
-        await expect(request).rejects.toMatchObject({ status });
-        expect(disclosedAPIKeys).toEqual([]);
-      } finally {
-        await Promise.all(
-          [source, destination].map(
-            (server) =>
-              new Promise<void>((resolve) => {
-                server.close(() => resolve());
-                server.closeAllConnections();
-              }),
-          ),
-        );
-      }
-    },
-  );
 });
 
 describe('azure request building', () => {
